requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
bleach==6.2.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.1.6
django-celery-results==2.5.1
djangorestframework==3.15.2
drf-api-logger==1.1.16
drf-yasg==1.21.8
exceptiongroup==1.2.2
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
